requirements/base.txt
Django==5.1.3
django-cors-headers==4.6.0
django-csp==3.8
django-environ==0.11.2
django-redis==5.4.0

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

psycopg2-binary==2.9.10

pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0

celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

requests==2.32.3